+++ pex/pex_bootstrapper.py
@@ -26,6 +26,7 @@ if TYPE_CHECKING:
         List,
         NoReturn,
         Optional,
+        Set,
         Tuple,
         Union,
     )
@@ -73,7 +74,7 @@ def iter_compatible_interpreters(
 
     def _iter_interpreters():
         # type: () -> Iterator[InterpreterOrError]
-        seen = set()
+        seen = set()  # type: Set[InterpreterOrError]
 
         normalized_paths = (
             OrderedSet(PythonInterpreter.canonicalize_path(p) for p in path.split(os.pathsep))
@@ -369,6 +370,11 @@ def ensure_venv(pex):
         raise AssertionError(
             "Expected PEX-INFO for {} to have the components of a venv directory".format(pex.path())
         )
+    if not pex_info.includes_tools:
+        raise ValueError(
+            "The PEX_VENV environment variable was set, but this PEX was not built with venv "
+            "support (Re-build the PEX file with `pex --venv ...`)"
+        )
     with atomic_directory(venv_dir, exclusive=True) as venv:
         if not venv.is_finalized:
             from .tools.commands.venv import populate_venv_with_pex
@@ -456,7 +462,7 @@ def bootstrap_pex(entry_point):
     # ENV.PEX_ROOT is consulted by PythonInterpreter and Platform so set that up as early as
     # possible in the run.
     with ENV.patch(PEX_ROOT=pex_info.pex_root):
-        if not ENV.PEX_TOOLS and pex_info.venv:
+        if not (ENV.PEX_UNZIP or ENV.PEX_TOOLS) and pex_info.venv:
             try:
                 target = find_compatible_interpreter(
                     interpreter_constraints=pex_info.interpreter_constraints,
@@ -465,7 +471,10 @@ def bootstrap_pex(entry_point):
                 die(str(e))
             from . import pex
 
-            venv_pex = ensure_venv(pex.PEX(entry_point, interpreter=target))
+            try:
+                venv_pex = ensure_venv(pex.PEX(entry_point, interpreter=target))
+            except ValueError as e:
+                die(str(e))
             os.execv(venv_pex, [venv_pex] + sys.argv[1:])
         else:
             maybe_reexec_pex(pex_info.interpreter_constraints)